Excel generation 1st draft.

+++ com.ibm.js.team.scm.supporttools/src/main/java/com/ibm/js/team/supporttools/scm/statistics/sizerange/IRangeCalculator.java
@@ -3,7 +3,7 @@ package com.ibm.js.team.supporttools.scm.statistics.sizerange;
 public interface IRangeCalculator {
 	
 	int getInterval(long size);
-	int getTopThreshold(int interval);
+	double getTopThreshold(int interval);
 	String getName();
 
 }
